bugfix: fix overflow for rate limit

+++ dfget/core/uploader/uploader_helper_test.go
@@ -57,7 +57,7 @@ func newTestPeerServer(workHome string) (srv *peerServer) {
 	cfg := helper.CreateConfig(nil, workHome)
 	srv = newPeerServer(cfg, 0)
 	srv.totalLimitRate = 1000
-	srv.rateLimiter = util.NewRateLimiter(int32(defaultRateLimit), 2)
+	srv.rateLimiter = util.NewRateLimiter(int64(defaultRateLimit), 2)
 	return srv
 }
